refact(domain-page): users
