+++ app/dashboard/src/pages/users/modals/reset-all-usage-modal.tsx
@@ -1,6 +1,6 @@
-import { DeleteIcon } from 'components/Dialog/Icons';
 import {
   Button,
+  chakra,
   Modal,
   ModalBody,
   ModalCloseButton,
@@ -13,63 +13,75 @@ import {
   useToast,
 } from '@chakra-ui/react';
 import { FC, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
-import { Icon } from './Icon';
+import { DocumentMinusIcon } from '@heroicons/react/24/outline';
+import { Icon } from 'components/Icon';
+import { useTranslation } from 'react-i18next';
 import { useUsers } from 'contexts/UsersContext';
 
-export type DeleteUserModalProps = {
-  deleteCallback?: () => void;
-};
+export const ResetIcon = chakra(DocumentMinusIcon, {
+  baseStyle: {
+    w: 5,
+    h: 5,
+  },
+});
 
-export const DeleteUserModal: FC<DeleteUserModalProps> = () => {
+type DeleteUserModalProps = {};
+
+export const ResetAllUsageModal: FC<DeleteUserModalProps> = () => {
   const [loading, setLoading] = useState(false);
-  const { deletingUser: user, onDeletingUser, deleteUser } = useUsers();
+  const { isResetingAllUsage, onResetAllUsage, resetAllUsage } = useUsers();
   const { t } = useTranslation();
   const toast = useToast();
   const onClose = () => {
-    onDeletingUser(null);
+    onResetAllUsage(false);
   };
-  const onDelete = () => {
-    if (user) {
-      setLoading(true);
-      deleteUser(user)
-        .then(() => {
-          toast({
-            title: t('deleteUser.deleteSuccess', { username: user.username }),
-            status: 'success',
-            isClosable: true,
-            position: 'top',
-            duration: 3000,
-          });
-        })
-        .then(onClose)
-        .finally(setLoading.bind(null, false));
-    }
+  const onReset = () => {
+    setLoading(true);
+    resetAllUsage()
+      .then(() => {
+        toast({
+          title: t('resetAllUsage.success'),
+          status: 'success',
+          isClosable: true,
+          position: 'top',
+          duration: 3000,
+        });
+      })
+      .catch(() => {
+        toast({
+          title: t('resetAllUsage.error'),
+          status: 'error',
+          isClosable: true,
+          position: 'top',
+          duration: 3000,
+        });
+      })
+      .finally(() => {
+        setLoading(false);
+      });
   };
   return (
-    <Modal isCentered isOpen={!!user} onClose={onClose} size="sm">
+    <Modal isCentered isOpen={isResetingAllUsage} onClose={onClose} size="sm">
       <ModalOverlay bg="blackAlpha.300" backdropFilter="blur(10px)" />
       <ModalContent mx="3">
         <ModalHeader pt={6}>
           <Icon color="red">
-            <DeleteIcon />
+            <ResetIcon />
           </Icon>
         </ModalHeader>
         <ModalCloseButton mt={3} />
         <ModalBody>
           <Text fontWeight="semibold" fontSize="lg">
-            {t('deleteUser.title')}
+            {t('resetAllUsage.title')}
           </Text>
-          {user && (
+          {isResetingAllUsage && (
             <Text
               mt={1}
               fontSize="sm"
               _dark={{ color: 'gray.400' }}
               color="gray.600"
             >
-              <Trans components={{ b: <b /> }}>
-                {t('deleteUser.prompt', { username: user.username })}
-              </Trans>
+              {t('resetAllUsage.prompt')}
             </Text>
           )}
         </ModalBody>
@@ -81,10 +93,10 @@ export const DeleteUserModal: FC<DeleteUserModalProps> = () => {
             size="sm"
             w="full"
             colorScheme="red"
-            onClick={onDelete}
+            onClick={onReset}
             leftIcon={loading ? <Spinner size="xs" /> : undefined}
           >
-            {t('delete')}
+            {t('reset')}
           </Button>
         </ModalFooter>
       </ModalContent>
